Bugfixes: count filtered. properly start calculating idle time.
The "filtered" count was wrong, add proper counting code.

The idle time code didn't account for the fact that we don't mount
/proc ourselves, and thus the first samples we get may be skipped,
resulting in no actual CPU numbers being present (0.0 in the data),
which equals idle. The proper fix is to not start looking for idle
time in the data, which is at the point that bootchart itself sees
its own first real data point.