Kernel init thread plotting: initcall_debug=1 data from dmesg put into a nice plot.
This addition basically replaces Arjan van der Ven's 'bootgraph.pl' script which can
be found in the scripts/ folder of the kernel sources. This implementation is highly
simplified and intended to show the basic relations between kernel threads taking
time to initialize.

To enable the initcall graph, just add 'initcall_debug=1' on the kernel commandline
additionally to the commandline code needed to start bootchartd.