Rewrite the main data structure to be dynamically allocated.
The 'ps' data array used to be a flat, stack allocated array of
65k entries of each 200k each. This makes no sense anymore, and
just puts us at the wrong end of VM tricks.

Instead, we malloc() a per-process struct ps_struct of about
200k whenever needed.

Because we now don't have a linear array anymore, we need to do
pointer juggling to walk the entries as an array (->next_ps), and
as a process tree (->childred, ->next). This makes walking and
maintaining the tree of processes a bit more complex.

All in all however the code cleans up a bit as we're now just
walking linked lists, one way or another.