Allow fractional logging frequencies.
This patch allows you to specify a float value for hz. The net
effect is that you can increase the logging interval to values
larger than 1.0 seconds, which is useful if you want to use
the profiling code to profile a larger timeframe.

I tested this with 0.1, 0.5, 22.5 values, and seems to work
as expected.