Signals: HUP -> write graph and exit.
I'm removing the TERM/INT signal handling, as I want those
to just terminate the application right away.

HUP is a more suited signal name for tricking the logger into
writing the graph out, so change the sighandler to that.

Eventually, I want to add a USR1 handler that causes bootchart
to plot and reset all it's data, but that's not a small patch.